+++ Sources/ImageOptions.swift
@@ -8,6 +8,12 @@
 
 import Foundation
 import CoreGraphics
+import ObjectMapper
+#if os(iOS) || os(tvOS) || os(watchOS)
+    import UIKit
+#elseif os(macOS)
+    import Cocoa
+#endif
 
 extension Asset {
 
@@ -226,7 +232,18 @@ public enum Focus: String {
  */
 public enum Fit: URLImageQueryExtendable {
 
-    case pad(withBackgroundColor: CGColor?)
+    #if os(iOS) || os(tvOS) || os(watchOS)
+    public typealias Color = UIColor
+    #else
+    public typealias Color = NSColor
+    #endif
+
+    /** 
+     If specifying an optional `UIColor` or `NSColor` make sure to also provide a custom width and height
+     or else you may receive an error from the server. If the color cannot be resolved to a hex string by the SDK,
+     an error will be thrown.
+     */
+    case pad(withBackgroundColor: Color?)
     case crop(focusingOn: Focus?)
     case fill(focusingOn: Focus?)
     case thumb(focusingOn: Focus?)
@@ -248,10 +265,14 @@ public enum Fit: URLImageQueryExtendable {
         return ImageParameters.fit
     }
 
-    fileprivate func additionalQueryItem() -> URLQueryItem? {
+    fileprivate func additionalQueryItem() throws -> URLQueryItem? {
         switch self {
         case .pad(let .some(color)):
-            return URLQueryItem(name: ImageParameters.backgroundColor, value: color.hexRepresentation())
+            let hexTransform = ObjectMapper.HexColorTransform()
+            guard let hexRepresentation = hexTransform.transformToJSON(color) else {
+                throw SDKError.invalidImageParameters("Unable to generate Hex representation for color: \(color)")
+            }
+            return URLQueryItem(name: ImageParameters.backgroundColor, value: "rgb:" + hexRepresentation)
 
         case .thumb(let .some(focus)):
             return URLQueryItem(name: ImageParameters.focus, value: focus.rawValue)
@@ -307,37 +328,4 @@ fileprivate struct ImageParameters {
     static let format           = "fm"
     static let quality          = "q"
     static let progressiveJPG   = "fl"
-}
-
-// Use CGColor instead of UIColor to enable cross-platform compatibility: macOS, iOS, tvOS, watchOS.
-internal extension CGColor {
-
-    // If for some reason the following code fails to create a hex string, the color black will be
-    // returned.
-    internal func hexRepresentation() -> String {
-        let hexForBlack = "000000"
-        guard let colorComponents = components else { return hexForBlack }
-        guard let colorSpace = colorSpace else { return hexForBlack }
-
-        let r, g, b: Float
-
-        switch colorSpace.model {
-        case .monochrome:
-            r = Float(colorComponents[0])
-            g = Float(colorComponents[0])
-            b = Float(colorComponents[0])
-
-        case .rgb:
-            r = Float(colorComponents[0])
-            g = Float(colorComponents[1])
-            b = Float(colorComponents[2])
-        default:
-            return hexForBlack
-        }
-
-        // Search the web for Swift UIColor to hex.
-        // This answer helped: https://stackoverflow.com/a/30967091/4068264
-        let hexString = String(format: "%02lX%02lX%02lX", lroundf(r * 255), lroundf(g * 255), lroundf(b * 255))
-        return hexString
-    }
 }