+++ CHANGELOG.md
@@ -10,17 +10,25 @@ This project adheres to [Semantic Versioning](http://semver.org/) starting from 
 - Add `@objc` annotations to capable methods for improved interoperability with Objective-C codebases.
 
 ### Merged, but not yet released
-> 
-
+>
 ---
 
 ## Table of contents
 
 #### 0.x Releases
-- `0.6.x` Releases - [0.6.0](#060)
+- `0.6.x` Releases - [0.6.0](#060) | [0.6.1](#061)
 - `0.5.x` Releases - [0.5.0](#050)
 - `0.4.x` Releases - [0.4.0](#040) | [0.4.1](#041)
 - `0.3.x` Releases - [0.3.0](#030) | [0.3.1](#031)
+
+---
+
+## [`0.6.1`](https://github.com/contentful/contentful.swift/releases/tag/0.6.1)
+Released on 2017-06-12
+
+#### Fixed
+- `ImageOption` that changed background using `Fit.pad(with: Color) now generates the correct URL.
+- Build error for watchOS caused by file from test target being added to watch target.
 
 ---